+++ src/constants/path.js
@@ -104,6 +104,7 @@ const API_URL = {
   FOLLOW_MATE: `/api/${VERSION}/${API_BASE.MATE}/search`,
   UNFOLLOW_MATE: `/api/${VERSION}/${API_BASE.MATE}/search`,
   EXPLORE_RANDOM_MATE: `/api/${VERSION}/${API_BASE.MATE}/random`,
+  LOCATION_MATE: `/api/${VERSION}/${API_BASE.MATE}/location`,
   GET_MATE_FOLLOWER: `/api/${VERSION}/${API_BASE.MATE}/followList`,
   GET_MATE_FOLLOWING: `/api/${VERSION}/${API_BASE.MATE}/followingList`,
   CREATE_MATE_RULE: `/api/${VERSION}/${API_BASE.MATE}/rule/write`,
