+++ src/hooks/mate/useGetLocationMate.js
@@ -1,0 +1,30 @@
+import { useEffect, useState } from 'react';
+
+import { getLocationMate } from '@/apis/request/mate';
+
+export const useGetLocationMate = () => {
+  const [data, setData] = useState([]);
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState(false);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        setLoading(true);
+        const res = await getLocationMate();
+        console.log(res);
+        const data = res.data.data;
+        console.log(data);
+        setData(data);
+      } catch (e) {
+        setError(e.message || '에러가 발생했습니다.');
+      } finally {
+        setLoading(false);
+      }
+    };
+
+    fetchData();
+  }, []);
+
+  return { data, loading, error };
+};
